+++ monocle/gerrit/review.py
@@ -159,6 +159,7 @@ class ReviewesFetcher(BaseCrawler):
                     "target_branch": change["target_branch"],
                     "number": change["number"],
                     "change_id": change["change_id"],
+                    "url": change["url"],
                     "on_author": change["author"],
                     "on_created_at": change["created_at"],
                     "changed_files": [